requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
daphne==2.4.1
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.8.1
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
ecdsa==0.15
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
rsa==4.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
whitenoise==5.0.1
zope.interface==4.7.2
